reorganize test folder
